Move fullName to a global function so it can be used by app
